Move head tracking from NpcAnimation to Animation (Bug #2720)
